% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANCOVA_analytic.R
\name{ANCOVA_analytic}
\alias{ANCOVA_analytic}
\title{Power Calculations for a factorial ANCOVA}
\usage{
ANCOVA_analytic(
  design,
  mu,
  n = NULL,
  sd,
  r2 = NULL,
  n_cov,
  alpha_level = Superpower_options("alpha_level"),
  beta_level = NULL,
  label_list = NULL,
  design_result = NULL,
  round_up = TRUE
)
}
\value{

}
\description{
Compute power of ANCOVA omnibus tests
}
\section{References}{

Keppel, G. (1991). Design and Analysis A Researcher's Handbook. 3rd Edition. Prentice Hall. Englewood Cliffs, New Jersey. See pages 323 - 324. 
Shieh, G. (2017). Power and sample size calculations for contrast analysis in ANCOVA. Multivariate behavioral research, 52(1), 1-11.
Shieh, G. (2020). Power analysis and sample size planning in ANCOVA designs. Psychometrika, 85(1), 101-120.
}

\examples{
# To be added
}
