% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.ancova_power.R
\name{ancova_power-methods}
\alias{ancova_power-methods}
\alias{print.ancova_power}
\alias{plot.ancova_power}
\title{Methods for ancova_power objects}
\usage{
\method{print}{ancova_power}(x, ...)

\method{plot}{ancova_power}(x, ...)
}
\arguments{
\item{x}{object of class \code{ancova_power} as returned from one of the simulation functions in Superpower.}

\item{...}{further arguments passed through, see description of return value}
}
\value{
\describe{
  \item{\code{print}}{Prints short summary of the simulation result}
  \item{\code{plot}}{Returns a meansplot of from the defined design}
}
}
\description{
Methods defined for objects returned from the ANCOVA_analytic function.
}
