% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA_design.R
\name{ANOVA_design}
\alias{ANOVA_design}
\title{Design function used to specify the parameters used in the simulation}
\usage{
ANOVA_design(design, n, mu, sd, r = 0, labelnames = NULL,
  plot = TRUE)
}
\arguments{
\item{design}{String specifying the ANOVA design.}

\item{n}{Sample size in each condition}

\item{mu}{Vector specifying mean for each condition}

\item{sd}{standard deviation for all conditions}

\item{r}{Correlation between dependent variables (single value or matrix)}

\item{labelnames}{Optional vector to specifying factor and condition names (recommended, if not used factors and levels are indicated by letters and numbers)}

\item{plot}{Should means plot be printed (defaults to TRUE)}
}
\value{
Returns Single data-frame with simulated data, design, design list, factor names, formulas for ANOVA, means, sd, correlation, sample size per condition, correlation matrix, covariance matrix, design string, labelnames, labelnameslist, factor names, meansplot
}
\description{
Design function used to specify the parameters used in the simulation
}
\section{References}{

too be added
}

\examples{
## Set up a within design with 2 factors, each with 2 levels,
## with correlation between observations of 0.8,
## 40 participants (who do all conditions), and standard deviation of 2
## with a mean pattern of 1, 0, 1, 0, conditions labeled 'condition' and
## 'voice', with names for levels of "cheerful", "sad", and "human", "robot"
ANOVA_design(design = "2w*2w", n = 40, mu = c(1, 0, 1, 0), sd = 2, r = 0.8,
      labelnames = c("condition", "cheerful", "sad", "voice", "human", "robot"))
}
