% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA_exact.R
\name{ANOVA_exact}
\alias{ANOVA_exact}
\title{Simulates on exact empirical data set from the design to calculate power}
\usage{
ANOVA_exact(design_result, correction = "none", alpha_level,
  verbose = TRUE)
}
\arguments{
\item{design_result}{Output from the ANOVA_design function}

\item{correction}{Set a correction of violations of sphericity. This can be set to "none", "GG" Grennhouse-Geisser, and "HF" Huynh-Feldt}

\item{alpha_level}{Alpha level used to determine statistical significance}

\item{verbose}{Set to FALSE to not print results (default = TRUE)}
}
\value{
Returns dataframe with simulation data (power and effect sizes), anova results and simple effect results, plot of exact data, and alpha_level.
}
\description{
Simulates on exact empirical data set from the design to calculate power
}
\section{References}{

to be added
}

\examples{
## Set up a within design with 2 factors, each with 2 levels,
## with correlation between observations of 0.8,
## 40 participants (who do all conditions), and standard deviation of 2
## with a mean pattern of 1, 0, 1, 0, conditions labeled 'condition' and
## 'voice', with names for levels of "cheerful", "sad", amd "human", "robot"
design_result <- ANOVA_design(design = "2w*2w", n = 40, mu = c(1, 0, 1, 0),
      sd = 2, r = 0.8, labelnames = c("condition", "cheerful",
      "sad", "voice", "human", "robot"))
 set.seed(252)
exact_result <- ANOVA_exact(design_result, alpha_level = 0.05)
}
