% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA_power.R
\name{ANOVA_power}
\alias{ANOVA_power}
\title{Simulation function used to perform the simulation}
\usage{
ANOVA_power(design_result, alpha_level = 0.05, correction = "none",
  p_adjust = "none", nsims = 1000, seed = NULL, verbose = TRUE)
}
\arguments{
\item{design_result}{Output from the ANOVA_design function}

\item{alpha_level}{Alpha level used to determine statistical significance}

\item{correction}{Set a correction of violations of sphericity. This can be set to "none", "GG" Grennhouse-Geisser, and "HF" Huynh-Feldt}

\item{p_adjust}{Correction for multiple comparisons}

\item{nsims}{number of simulations to perform}

\item{seed}{Set seed for reproducible results}

\item{verbose}{Set to FALSE to not print results (default = TRUE)}
}
\value{
Returns dataframe with simulation data (p-values and effect sizes), anova results and simple effect results, plots of p-value distribution, p_adjust = p_adjust, nsims, and alpha_level.
}
\description{
Simulation function used to perform the simulation
}
\section{References}{

too be added
}

\examples{
## Set up a within design with 2 factors, each with 2 levels,
## with correlation between observations of 0.8,
## 40 participants (who do all conditions), and standard deviation of 2
## with a mean pattern of 1, 0, 1, 0, conditions labeled 'condition' and
## 'voice', with names for levels of "cheerful", "sad", amd "human", "robot"
design_result <- ANOVA_design(design = "2w*2w", n = 40, mu = c(1, 0, 1, 0),
      sd = 2, r = 0.8, labelnames = c("condition", "cheerful",
      "sad", "voice", "human", "robot"))
power_result <- ANOVA_power(design_result, alpha_level = 0.05,
      p_adjust = "none", seed = 2019, nsims = 10)
}
