% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.design_aov.R
\name{design_aov-methods}
\alias{design_aov-methods}
\alias{print.design_aov}
\alias{plot.design_aov}
\title{Methods for design_aov objects}
\usage{
\method{print}{design_aov}(x, ...)

\method{plot}{design_aov}(x, ...)
}
\arguments{
\item{x}{object of class \code{design_aov} as returned from \code{ANOVA_design}}

\item{...}{further arguments passed through, see description of return value
for details.
\code{\link{ANOVA_design}}.}
}
\value{
\describe{
  \item{\code{print}}{Prints short summary of the study design created from \code{ANOVA_design} function}
  \item{\code{plot}}{Returns \code{meansplot} from created from the \code{ANOVA_design} function}
}
}
\description{
Methods defined for objects returned from the ANOVA_design functions.
}
