% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_from_design.R
\name{mu_from_design}
\alias{mu_from_design}
\title{Get means from Cohen's f}
\usage{
mu_from_design(x, f)
}
\value{

}
\description{
A function for producing a pattern of means that will
}
\examples{
# To be added
}
