% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.opt_alpha.R
\name{opt_alpha-methods}
\alias{opt_alpha-methods}
\alias{print.opt_alpha}
\alias{plot.opt_alpha}
\title{Methods for opt_alpha objects}
\usage{
\method{print}{opt_alpha}(x, ...)

\method{plot}{opt_alpha}(x, ...)
}
\arguments{
\item{x}{object of class \code{opt_alpha} as returned from one of the optimal alpha functions in Superpower.}

\item{...}{further arguments passed through, see description of return value
for details.
\code{\link{ANOVA_compromise}}.}
}
\value{
\describe{
  \item{\code{print}}{Prints short summary of the optimal alpha results}
  \item{\code{plot}}{Returns a plot}
}
}
\description{
Methods defined for objects returned from the optimal_alpha and ANOVA_compromise functions.
}
