% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_power.R
\name{plot_power}
\alias{plot_power}
\title{Convenience function to plot power across a range of sample sizes.}
\usage{
plot_power(design_result, alpha_level, min_n = 7, max_n = 100,
  plot = FALSE)
}
\arguments{
\item{design_result}{Output from the ANOVA_design function}

\item{alpha_level}{Alpha level used to determine statistical significance}

\item{min_n}{Minimum sample size in power curve.}

\item{max_n}{Maximum sample size in power curve.}

\item{plot}{Should power plot be printed automatically (defaults to FALSE)}
}
\value{
Returns plot with power curves for the ANOVA, and a dataframe with the summary data.

\describe{
  \item{\code{"plot_ANOVA"}}{Plot of power curves from ANOVA results.}
  \item{\code{"plot_MANOVA"}}{Plot of power curves from MANOVA results. Returns NULL if no within-subject factors.}
  \item{\code{"power_df"}}{The tabulated ANOVA power results.}
  \item{\code{"power_df_manova"}}{The tabulated MANOVA power results..}
  \item{\code{"effect_sizes"}}{Effect sizes (partial eta-squared) from ANOVA results.}
  \item{\code{"effect_sizes_manova"}}{Effect sizes (Pillai's Trace) from MANOVA results. Returns NULL if no within-subject factors.}
  
}
}
\description{
Convenience function to plot power across a range of sample sizes.
}
\section{References}{

too be added
}

\examples{
design_result <- ANOVA_design(design = "3b",
                             n = 20,
                             mu = c(0,0,0.3),
                             sd = 1,
                             labelnames = c("condition",
                             "cheerful", "neutral", "sad"))

plot_power(design_result, min_n = 50, max_n = 70)
}
