% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANCOVA_analytic.R
\name{ANCOVA_analytic}
\alias{ANCOVA_analytic}
\title{Power Calculations for a factorial ANCOVA}
\usage{
ANCOVA_analytic(
  design,
  mu,
  n = NULL,
  sd,
  r2 = NULL,
  n_cov,
  alpha_level = Superpower_options("alpha_level"),
  beta_level = NULL,
  contrasts = list(),
  label_list = NULL,
  design_result = NULL,
  round_up = TRUE
)
}
\value{

}
\description{
Compute power of ANCOVA omnibus tests and contrasts.
}
\section{References}{

Shieh, G. (2020). Power analysis and sample size planning in ANCOVA designs. Psychometrika, 85(1), 101-120.
}

\examples{
# To be added
}
