% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmvpower.h.R
\name{jmvpower}
\alias{jmvpower}
\title{ANOVA power}
\usage{
jmvpower(design = "2b*2w",
  labelnames = "AGE,old,young,SPEED,fast,slow", n = 80, sd = 1.03,
  r = 0.87, mu = "1.03, 1.21, 0.98, 1.01", plot = TRUE,
  simulate = FALSE, alpha_level = 0.05, nsims = 100,
  p_adjust = "holm")
}
\arguments{
\item{design}{a string describing the design}

\item{labelnames}{a comma separated string describing the factor and level
labels}

\item{n}{an integer specifying the sample size in each condition}

\item{sd}{a number specifying the group standard deviations}

\item{r}{a number specifying the correlation between dependent variables}

\item{mu}{a comma separated string specifying the group means}

\item{plot}{\code{TRUE} (default) or \code{FALSE} specifying whether to
provide a means plot}

\item{simulate}{\code{TRUE} or \code{FALSE} (default); perform the
simulation}

\item{alpha_level}{a number specifying the alpha level}

\item{nsims}{an integer specifying the number of simulations to perform}

\item{p_adjust}{the p-adjustment method to use}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$design$summary} \tab \tab \tab \tab \tab a table \cr
  \code{results$design$matrix} \tab \tab \tab \tab \tab a table \cr
  \code{results$design$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$sims$results} \tab \tab \tab \tab \tab a table \cr
  \code{results$sims$multi} \tab \tab \tab \tab \tab a table \cr
}
}
\description{
ANOVA power
}
