% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_power.R
\name{plot_power}
\alias{plot_power}
\title{Convenience function to plot power across a range of sample sizes.}
\usage{
plot_power(design_result, alpha_level, min_n = 7, max_n = 100,
  plot = TRUE)
}
\arguments{
\item{design_result}{Output from the ANOVA_design function}

\item{alpha_level}{Alpha level used to determine statistical significance}

\item{min_n}{Minimum sample size in power curve.}

\item{max_n}{Maximum sample size in power curve.}

\item{plot}{Should power plot be printed (defaults to TRUE)}
}
\value{
Returns plot with power curves for the ANOVA, and a dataframe with the summary data.
}
\description{
Convenience function to plot power across a range of sample sizes.
}
\section{References}{

too be added
}

\examples{
design_result <- ANOVA_design(design = "3b",
                             n = 20,
                             mu = c(0,0,0.3),
                             sd = 1,
                             labelnames = c("condition",
                             "cheerful", "neutral", "sad"))

plot_power(design_result, min_n = 50, max_n = 70)
}
