% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.sim_result.R
\name{sim_result-methods}
\alias{sim_result-methods}
\alias{print.sim_result}
\alias{plot.sim_result}
\title{Methods for sim_result objects}
\usage{
\method{print}{sim_result}(x, ...)

\method{plot}{sim_result}(x, ...)
}
\arguments{
\item{x}{object of class \code{sim_result} as returned from one of the simulation functions in Superpower}

\item{...}{further arguments passed through, see description of return value
for details.
\code{\link{ANOVA_design}}.}
}
\value{
\describe{
  \item{\code{print}}{Prints short summary of the simulation result}
  \item{\code{plot}}{Returns \code{meansplot} or a plot of the distribution of p-values depending on whether an exact or Monte Carlo simulation was performed}
}
}
\description{
Methods defined for objects returned from the ANOVA_exact, ANOVA_exact2, and ANOVA_power functions.
}
