% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_ANCOVA.R
\name{simple_ANCOVA}
\alias{simple_ANCOVA}
\title{Monte carlo simulation function for simple pre-post parellel group design}
\usage{
simple_ANCOVA(post_diff, sd = 1, r = 0.5, min_n = 7, max_n = 100,
  alpha_level = 0.05, conf_level, nsims = 1000, MOE)
}
\arguments{
\item{sd}{standard deviation for all conditions (i.e., common standard deviation; default is 1)}

\item{r}{Correlation between pre and post measurement (default is .5)}

\item{min_n}{Minimum sample size in for the simulations.}

\item{max_n}{Maximum sample size in for the simulations.}

\item{alpha_level}{Alpha level used to determine statistical significance}

\item{conf_level}{The confidence (or compatability) level. Default is 1-alpha_level.}

\item{nsims}{number of simulations to perform at each sample size. Default is 1000.

\describe{
  \item{\code{"sim_results"}}{Power and precision of both statistical tests at the various sample sizes}
  \item{\code{"plot_power_ANCOVA"}}{Plot of the power for ANCOVA results}
  \item{\code{"plot_power_ANOVA"}}{Plot of the power for ANOVA results}
  \item{\code{"plot_precision_ANCOVA"}}{Plot of assurance of the confidence interval for the ANCOVA estimate being within the margin of error}
  \item{\code{"plot_precision_ANOVA"}}{Plot of assurance of the confidence interval for the ANOVA estimate being within the margin of error}

}}

\item{MOE}{The margin of error. The desired length of one arm of a CI, or the maximum likely estimation error. (see Cummings and Calin-Jageman)}

\item{diff_post}{Difference between groups at the post time (defalt is zero)}
}
\value{
Returns dataframe with simulation results at each sample size as well as plots of the confidence intervals widths and power for the ANOVA and ANCOVA result.
}
\description{
Monte carlo simulation function for simple pre-post parellel group design
}
\section{References}{

Cumming, Geoff, and Robert Calin-Jageman. Introduction to the new statistics: Estimation, open science, and beyond. Routledge, 2016.
}

\examples{
Too be added
}
